Тесты компонентов MovieList и SmallMovieCard
